+++ counterpartylib/lib/transaction.py
@@ -391,8 +391,8 @@ def construct (db, tx_info, encoding='auto',
     outputs_size = ((25 + 9) * len(destination_outputs)) + (len(data_array) * data_output_size)
 
     # Get inputs.
-    unspent = backend.get_unspent_txouts(source)
-    unspent = backend.sort_unspent_txouts(unspent, allow_unconfirmed_inputs)
+    unspent = backend.get_unspent_txouts(source, unconfirmed=allow_unconfirmed_inputs)
+    unspent = backend.sort_unspent_txouts(unspent)
     logger.debug('Sorted UTXOs: {}'.format([print_coin(coin) for coin in unspent]))
 
     inputs = []