+++ ChangeLog.md
@@ -1,6 +1,8 @@
 ## Client Versions ##
 * v9.49.5 (???)
 	* automatically spend 1‐of‐`N` multisig outputs
+	* remove `get_asset_info(assets)` method from the API
+	* replace `get_xcp_supply()` method from the API by `get_supply(asset)`
 * v9.49.4 (2014-02-05)
 	* reconceived this package as a libary
 	* moved CLI to new repository: `counterparty-cli`
@@ -18,6 +20,7 @@
 	* improved docstring coverage
 	* removed option to use Insight, Blockr and SoChain as a backend
 	* tweaked Coveralls configuration
+	* Rename the database file name from `counterpartyd.9.db` to `counterparty.db`
 	* Add BTCD support
 * v9.49.3 (2014-12-28)
 	* better logging when handling `SIGTERM`, `SIGKILL`