+++ ChangeLog.md
@@ -3,6 +3,7 @@
 	* automatically spend 1‐of‐`N` multisig outputs
 	* remove `get_asset_info(assets)` method from the API
 	* replace `get_xcp_supply()` method from the API by `get_supply(asset)`
+	* added HTTP REST API
 * v9.49.4 (2014-02-05)
 	* reconceived this package as a libary
 	* moved CLI to new repository: `counterparty-cli`