+++ ChangeLog.md
@@ -7,6 +7,7 @@
 	* added HTTP REST API
 	* authentication on JSON-RPC API is off by default
 	* RPC password is not mandatory now
+	* `get_unspent_txouts` API method parameter and return values changed
 * v9.49.4 (2014-02-05)
 	* reconceived this package as a libary
 	* moved CLI to new repository: `counterparty-cli`