+++ ChangeLog.md
@@ -1,4 +1,6 @@
 ## Client Versions ##
+* v9.50.0 (2014-03-18)
+	* hotfix: global integer overflow
 * v9.49.4 (2014-02-05)
 	* reconceived this package as a libary
 	* moved CLI to new repository: `counterparty-cli`